fix!(components): renamed my components
